Add files via upload
This is Icici Bank Management System project to analyze the bank performance. The existing manual processes within ICICI Bank lead to inefficiencies in operations, delays in customer service, and higher risks of errors. There is a need for an automated and integrated management system to streamline operations, improve customer service, and enhance overall efficiency